+++ Main.java
@@ -2,17 +2,18 @@ import java.io.IOException;
 import java.net.InetSocketAddress;
 
 import com.sun.net.httpserver.HttpServer;
+public class Main{
+    public static void main(String[] args) throws IOException{
+        
 
-public class Main {
-    public static void main(String[] args) throws IOException {
-        InetSocketAddress endereco = new InetSocketAddress(8888);
+        InetSocketAddress endereco = new InetSocketAddress(8889);
+        HttpServer servidor = HttpServer.create(endereco,10);
+        servidor.createContext("/multiplicacao", new Multiplicacao());
 
-        HttpServer servidor = HttpServer.create(endereco, 10);
-        servidor.createContext("/", new IndexHandler());
-        servidor.createContext("/ehchatinho", new TuQueEh());
         servidor.start();
 
-
-        System.out.println("Servidor iniciado ouvindo a port " + endereco.getPort());
+        System.out.println("Servidor outro iniciado em "+ endereco.getPort());
     }
 }
+
+
